fix: Refactor user hooks
