+++ src/hooks/useAuth.js
@@ -1,0 +1,44 @@
+import { useState } from "react";
+import { statusApi } from "../api/statusApi";
+
+export const useAuth = () => {
+    const [username, setUsername] = useState('');
+    const [password, setPassword] = useState('');
+
+    const handleUsernameChange = (event) => {
+        setUsername(event.target.value);
+    };
+
+    const handlePasswordChange = (event) => {
+        setPassword(event.target.value);
+    };
+
+    const handleSubmit = (event) => {
+        event.preventDefault();
+    
+        statusApi.post('http://localhost:3001/api/user/signIn', {
+            username: username,
+            password: password
+        }, {
+            headers: {
+                'Content-Type': 'application/json',
+            }
+        })
+        .then((response) => {
+            const now = new Date();
+            const oneHourLater = new Date(now.getTime()+ 60 * 60 * 1000);
+            const oneDayLater = new Date(now.getTime()+ 24 * 60 * 60 * 1000);
+            const accessExpires = oneHourLater.toUTCString();
+            const refreshExpires = oneDayLater.toUTCString();
+
+            document.cookie = `accessToken=${response.data.accessToken}; expires=${accessExpires}`;
+            document.cookie = `refreshToken=${response.data.refreshToken}; expires=${refreshExpires}`;
+            window.location.href = '/';
+        })
+        .catch((error) => {
+            console.error(error.message);
+        });
+    };
+
+    return { username, password, handleUsernameChange, handlePasswordChange, handleSubmit };
+};
