+++ src/pages/auth/Login.js
@@ -1,45 +1,9 @@
-import React, { useState } from 'react';
-import axios from 'axios';
+import React from 'react';
+import { useAuth } from '../../hooks/useAuth';
 
 
 export default function Login() {
-    const [username, setUsername] = useState('');
-    const [password, setPassword] = useState('');
-
-    const handleUsernameChange = (event) => {
-        setUsername(event.target.value);
-    };
-
-    const handlePasswordChange = (event) => {
-        setPassword(event.target.value);
-    };
-
-    const handleSubmit = (event) => {
-        event.preventDefault();
-    
-        axios.post('http://localhost:3001/api/user/signIn', {
-            username: username,
-            password: password
-        }, {
-            headers: {
-                'Content-Type': 'application/json',
-            }
-        })
-        .then((response) => {
-            const now = new Date();
-            const oneHourLater = new Date(now.getTime()+ 60 * 60 * 1000);
-            const oneDayLater = new Date(now.getTime()+ 24 * 60 * 60 * 1000);
-            const accessExpires = oneHourLater.toUTCString();
-            const refreshExpires = oneDayLater.toUTCString();
-
-            document.cookie = `accessToken=${response.data.accessToken}; expires=${accessExpires}`;
-            document.cookie = `refreshToken=${response.data.refreshToken}; expires=${refreshExpires}`;
-            window.location.href = '/';
-        })
-        .catch((error) => {
-            console.error(error.message);
-        });
-    };
+    const { username, password, handleUsernameChange, handlePasswordChange, handleSubmit } = useAuth();
     
     return (
         <div>
@@ -67,6 +31,4 @@ export default function Login() {
             </form>
         </div>
     );
-
-    
 }
